# crates/puddle-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "puddle",
    "description": "puddle's local API: pending requests, rules, audit, settings, consents and the event stream. Listens on 127.0.0.1 only. Every request needs `Authorization: Bearer <token>` (from the connection file), a `Host` of `127.0.0.1:<port>` or `localhost:<port>`, and no foreign `Origin`.",
    "contact": {
      "name": "Tijs van Kampen",
      "email": "puddle@tijsvankampen.be"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/inbox": {
      "get": {
        "tags": [
          "pending"
        ],
        "summary": "Open requests grouped by registrable domain, most recent group first (R-18).",
        "operationId": "inbox",
        "responses": {
          "200": {
            "description": "the inbox",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inbox"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pending": {
      "get": {
        "tags": [
          "pending"
        ],
        "summary": "Open (`requested`) requests, most recent first.",
        "operationId": "list_pending",
        "parameters": [
          {
            "name": "workspace",
            "in": "query",
            "description": "only this workspace's requests",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/WorkspaceName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "open requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingList"
                }
              }
            }
          },
          "400": {
            "description": "invalid workspace name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/pending/{id}": {
      "get": {
        "tags": [
          "pending"
        ],
        "summary": "One pending request, in any state.",
        "operationId": "get_pending",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "pending request id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingRequest"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/pending/{id}/approve": {
      "post": {
        "tags": [
          "pending"
        ],
        "summary": "Approves an open request: creates an allow rule (this workspace, exact host, permanent unless\nthe body says otherwise) and closes the other open requests it now decides.",
        "operationId": "approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "pending request id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOutcome"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "already decided or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "invalid suffix or expiry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/pending/{id}/deny": {
      "post": {
        "tags": [
          "pending"
        ],
        "summary": "Denies an open request: creates a deny rule (this workspace, exact host, permanent unless the\nbody says otherwise) and closes the other open requests it now decides.",
        "operationId": "deny",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "pending request id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOutcome"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "already decided or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "invalid suffix or expiry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspace}/suppression": {
      "get": {
        "tags": [
          "pending"
        ],
        "summary": "Whether a workspace's new pending requests are being suppressed (R-13), for the inbox's\n\"N requests suppressed\" line.",
        "operationId": "suppression",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "description": "workspace name",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkspaceName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the workspace's suppression state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suppression"
                }
              }
            }
          },
          "400": {
            "description": "invalid workspace name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Actor": {
        "type": "string",
        "description": "Who made a change.",
        "enum": [
          "cli",
          "ui",
          "api",
          "system"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, lower case, no trailing period. Never contains a token."
          }
        }
      },
...
      "DecisionOutcome": {
        "type": "object",
        "description": "What an approve or deny did (R-16, R-17).",
        "required": [
          "request",
          "rule",
          "also_closed"
        ],
        "properties": {
          "also_closed": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Other open requests the new rule closed the same way."
          },
          "request": {
            "$ref": "#/components/schemas/PendingRequest",
            "description": "The request, now decided."
          },
          "rule": {
            "$ref": "#/components/schemas/Rule",
            "description": "The rule created."
          }
        }
      },
      "DecisionRequest": {
        "type": "object",
        "description": "The choices of an approve or deny (R-15). Each one left out stays at its narrowest default:\nthis workspace, the exact host, permanent.",
        "properties": {
          "expires_in_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seconds until the rule expires; left out or `null` is permanent.",
            "minimum": 1
          },
          "rule_set": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Put the rule into this rule set of yours (the number of `user:<id>`) instead: it then\napplies wherever the set is on. `scope` must be left at `workspace`. The set must be on\nfor the request's workspace."
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeChoice",
            "description": "This workspace (default) or every workspace."
          },
          "suffix": {
            "type": [
              "string",
              "null"
            ],
            "description": "A suffix of the host (`example.com`, `.example.com` or `*.example.com`) to cover every\nname under it; left out or `null` means the exact host. A public suffix is refused."
          }
        },
        "additionalProperties": false
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "What went wrong, as a stable code a client can switch on.",
        "enum": [
          "misdirected_host",
          "forbidden_origin",
          "unauthorized",
          "bad_request",
          "unsupported_media_type",
          "payload_too_large",
          "invalid",
          "not_found",
          "conflict",
          "newer_settings",
          "unavailable",
          "internal"
        ]
      },
...
      "Inbox": {
        "type": "object",
        "description": "The inbox: open requests grouped by registrable domain, most recent group first.",
        "required": [
          "groups"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InboxGroup"
            },
            "description": "The groups."
          }
        }
      },
      "InboxGroup": {
        "type": "object",
        "description": "Open requests that share a registrable domain (R-18). Grouping is for display only.",
        "required": [
          "registrable_domain",
          "requests"
        ],
        "properties": {
          "registrable_domain": {
            "type": "string",
            "description": "`example.co.uk`, or the IP literal."
          },
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingRequest"
            },
            "description": "Open requests, most recent first."
          }
        }
      },
...
      "LocalCategory": {
        "type": "string",
        "description": "A class of local destination with its own toggle. Everything outside these is public.\n\nA toggle only *permits* its category; each destination still needs an allow rule or an\napproval. puddle's own endpoints are not a category: no toggle reaches them.\n\nNon-exhaustive: a \"company network\" category may follow.",
        "enum": [
          "loopback",
          "private",
          "link_local",
          "metadata",
          "special"
        ]
      },
...
      "PendingList": {
        "type": "object",
        "description": "Pending requests, most recent first.",
        "required": [
          "requests"
        ],
        "properties": {
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingRequest"
            },
            "description": "The requests."
          }
        }
      },
      "PendingRequest": {
        "type": "object",
        "description": "A connection that matched no rule and waits for the user (R-10).",
        "required": [
          "id",
          "workspace",
          "host",
          "port",
          "first_seen",
          "last_seen",
          "attempts",
          "state",
          "decided_at",
          "decided_by",
          "rule_id",
          "rule_set",
          "blocked_by"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int64",
            "description": "How many requests the row stands for.",
            "minimum": 0
          },
          "blocked_by": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LocalCategory",
                "description": "For an open request to a local destination (host loopback, private network, link-local,\nmetadata, special) whose toggle is off in this workspace: the toggle that blocks approval.\nAn allow rule would not let the connection through until the toggle is on. `null`\notherwise, and always for a closed request."
              },
              {
                "type": "null"
              }
            ]
          },
          "decided_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Epoch ms it was decided or expired.",
            "minimum": 0
          },
          "decided_by": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Actor",
                "description": "Who closed it."
              },
              {
                "type": "null"
              }
            ]
          },
          "first_seen": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch ms of the first request.",
            "minimum": 0
          },
          "host": {
            "type": "string",
            "description": "The requested host: a normalised name (punycode, lower case) or an IP literal. Untrusted\n(it comes from the guest): escape it when rendering."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Row id, never reused."
          },
          "last_seen": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch ms of the latest request.",
            "minimum": 0
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "description": "The requested port.",
            "minimum": 0
          },
          "rule_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The rule that decided it."
          },
          "rule_set": {
            "type": [
              "string",
              "null"
            ],
            "description": "The rule set whose entry decided it (`system`, `builtin:<slug>`, `user:<id>`), when one\ndid."
          },
          "state": {
            "$ref": "#/components/schemas/PendingState",
            "description": "Where the row is in its life."
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceName",
            "description": "The requesting workspace."
          }
        }
      },
      "PendingState": {
        "type": "string",
        "description": "Where a pending request is in its life (R-12).",
        "enum": [
          "requested",
          "allowed",
          "denied",
          "expired"
        ]
      },
...
      "ScopeChoice": {
        "type": "string",
        "description": "Whether an approval or denial covers only the request's workspace or every workspace.",
        "enum": [
          "workspace",
          "global"
        ]
      },
...
      "Suppression": {
        "type": "object",
        "description": "Whether a workspace's new pending requests are being suppressed (R-13).",
        "required": [
          "workspace",
          "active",
          "count"
        ],
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Whether requests are being suppressed now."
          },
          "count": {
            "type": "integer",
            "format": "int64",
            "description": "Requests suppressed in the current (or last) episode.",
            "minimum": 0
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceName",
            "description": "The workspace."
          }
        }
      },
...
      "WorkspaceName": {
        "type": "string",
        "description": "A workspace's name: a DNS label (`a-z`, `0-9`, `-`, no leading or trailing `-`), not `tauri`, `ipc` or `asset`.",
        "example": "my-project",
        "maxLength": 63,
        "minLength": 1,
        "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}